requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart
